Redesigned Save System, added GameData class
Completely redesigned the save system because there is some issue with the Godot saving nested custom resources (or at least I don't know how to make it work). Now Playbooks and SaveData save to the file system in an orderly way, and load as well.

I had to tear a lot of what was in the SaveGame resource and move it to the GameData, but it feels a lot cleaner now. GameData cleans up and stores a reference to all the resources being used, but itself shouldn't need to be accessed directly all that often. Mostly things can just hook into signals from GameSaver or GameData.

TODO:
Redo loading screen to account for changes (and be able to actually load games)
Reconnect Crew and Character Sheet Abilities to the appropriate data
Finish Crew and Character creation (and update data structures to account for the currently missing data)

after that ... clocks and map notes!