Major work on map notes
Fixed up a lot of random little bugs, added animation to map notes, in the middle of creating dialog for updating map notes once clicked. So far it looks like it should be smooth sailing hooking up the map notes into the save system and having them be placable by the user.