Notes working! 0.45
Got notes working, fixed a bunch of connection issues with various parts. I THINK everything currently in is working.

Also converted the notes into a grid instead of world position. Had to get rid of the minimap as the second viewport as it was creating duplicate map note enteries despite being set to not visible. WIll troubleshoot it later maybe, but honestly I may not need a minimap