Note/Location loading from SRD
Can now add notes to the SRD and load them properly when creating the map. It should work even if the srd is updated between loads.

Still need to add a way to move the notes and add an image and change the map icon.

Also deleting notes seems to be broken for some reason. need to look into that.