0.48 Major work on tooltip, fixed some bugs
Made behavior of UI elements better, did major work on the tooltip, now able to freeze the tooltip to tooltip your tooltips and so on. Tooltips all the way down.
Deleted some old code and nodes to clean up the codebase a little.

Learned that _ready() and _process() do not get overloaded when inherited. Good to know!

Need to add resizability to the DragableWindow
Eventually adding a custom text parser that can search the dictionary and provide in text links within the popups. Also want the ability to be able to open a contact or a location or whatever by clicking on it. It's a bit down the road though.