0.52 Still Working on Networking
Still working on getting networking to work, getting close but not quite there. Did some setup on getting rpcs to work, the goal is to get the players to update correctly and then make sure the GameData is synced. I'm not sure about transferring custom resources over the network with JSON, I think it likely won't work and I'll have to rework the setup functions in the playbooks. Ideally though when the playbooks get updated they send out their updates over the server and the other playbooks update. That should make all the fields update automatically. Addtionally there will have to be some setup for getting the initial play state when loading in from a server, and making sure that that isn't overriding any personal data.

Server-side Improvements to Make:
-Converting the server to athoratative so it can have labels and is able to store some things like default SRD data and Maps (maybe), but for sure I'd like to have labels and passwords on the server.
- For now I still plan to run the OP code logic on the game side
-In theory everything could be moved over to the server, allowing cloud saves, since updates to the gamedata have to be sent over the server anyway, I don't think it would take more network traffic to just save the changes on the server and send the updates to the client. Of course this would increase the storage requirements on the server, but it may not be that much considering all big files (like images) could be just stored once and referenced, and the individual playbooks are just some text files.