0.53 Networking Updates
Lots working now with multiplayer:
Player sprite and position
Notifications
Saveable Fields (mostly) update correctly on all machines
Markers update correct
Adding a character updates on all machines

Big thing is getting the inital game state that still needs to be implemented, when someone connects they need to download all the pc playbooks, the crew playbooks and the map and clocks from the host.

I should probably set it up so that the client requests from the host the game state.
The host tells the client the filenames of the saved playbooks and crew playbook
the client responds and says if they have or don't have the file
the host sends the file if needed
the host then just sends the map, map locations, and clocks to make sure those are all updated

This should cut down on network traffic if the playbooks are the same, and it shouldn't make a HUGE difference if they are SLIGHTLY different, as one update to the relevant field will update everyone to be on the same page anyway.