Reworking some things....
Put a bunch of functions in because Godot can't save resources with custom types well (really weird errors), so the custom types that are being saved in SaveGame (Clocks and Map Locations) now are converted into a dict of important values and can be loaded from the values (happens auto-matically when the game is saved and loaded from save)

The good thing is that the Nodes themselves are what is being stored in the GameData dict and can be referenced from anywhere...although I"m realizing that may be a bad thing considering they may need to be loaded from more than one node.

Yup I think I messed that up. Ugh, I'll refactor it later.