v0.57 Networked Resources
Changed courses hopefully for the last time, created NetworkedResources and a Library class to be able to catalog and search for them. Seems like it works really well so far, and the resources themselves are network enabled which means they'll update online. The thing I need to be careful about is making sure the ids don't get created in more than one place, but other than that they should work great.

Right now I'm using one OP_CODE for all networked resources, which in theory could mean some with the same ID would cause a bug, but there's about 1/1000000 chance of that happening. Still would be nice to make sure it doesn't happen.

Clocks are still on the old system, but working, I will move them over once I've got the other things hooked up and working.

The networked resources are amazing though and make everything much easier to work with. Any node can subscribe to the resource and is able to update it and will be informed when it gets updated. Plus, they transmit the same updates over the network.