v0.61 Clocks moved over!
Moved clocks over to the new resource system, began ripping out the code for GameData to update and recieve data. Now that is all done with the Library and the NetworkedResource classes. These classes store their data, check to see when it's updated, appropriately update the network if online, and let other things know that they have updated properties so they can update themselves accordingly as well.

System seems to work really well!

The advantage of using the Library class is that it knows when something is added and removed and can update the nodes that care. The GameData keeps it's data (and therefore SaveGame's data) updated based on the library calls and now no longer needs to care when things are updated in the game.