v0.62 Removing Playbooks
WAAAH my beloved class I was so proud of, now in trash. Moving everything over to NetworkedResource because it just works so much better. The Library - NetworkedResource schema is just fantastic for functionality and decoupling. GameData isn't responsible for updating everything, the resources all connect on their own. It also unhooks LOTS of nodes from having to connect to the network events, they just connect to their assigned resource, and when the resource is updated, or deleted, or whatever then it updates the node representation (and vica versa), as well the container nodes can hook into the libraries of resources and stay updated.

Also did another pass on find()ing data in the resources and made it so it's very easy to submit a path and get a result (and you can update on paths as well)
simple as "x.y.z.0" (yes it does array indexes as well)

I probably need to make some kind of list functionality of some kind.