v0.66 Working on getting everything functioning still...
Still working in getting everything up and running on the new NetworkedResources and Library system. So far it seems almost ready, but not quite to the point where I can change maps around and such.

The big issue has been saving and loading as little things keep going wrong, once I get that stable it should be a lot easier to move forward. Currently I'm not working on being able to REMOVE data, just to update, add, and change what is currently showing. Once all that works it should be pretty easy to add remove functionality - most the code is there already.

The way the map SHOULD work, is the GameData map variable is set to a mapID that shows the currently displayed map, the map_library should have all the available map resources in it,  but if you grab GameData.map it just gives you the NetworkedResource for the correct map.

Once it works it should be effortless to switch between maps, create a new map and save/load maps along with all the other save data. similarly all the map resources should update over the network because it's using the same NetworkedResources as everything else.